requirements.txt
asgiref==3.2.10
astroid==2.4.2
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.11.8
chardet==3.0.4
cycler==0.10.0
Django==3.1.2
djangorestframework==3.12.2
django-admin-rangefilter==0.6.3
django-advanced-filters==1.2.0
django-braces==1.13.0
django-taggit==1.3.0
djangoql==0.14.3
html5lib==1.1
idna==2.10
isort==5.6.4
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
logilab-astng==0.24.3
logilab-common==1.8.0
lxml==4.6.1
matplotlib==3.3.3
mccabe==0.6.1
mechanize==0.4.5
mypy-extensions==0.4.3
numpy==1.19.4
pandas==1.1.4
Pillow==8.0.1
ply==3.11
pycodestyle==2.6.0
pycountry==20.7.3
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytz==2020.1
requests==2.25.0
simplejson==3.17.2
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
webencodings==0.5.1
wrapt==1.12.1
mysqlclient==2.0.1
django-mysql==2.2.0